iam_portal/terraform/src/granting/requirements.txt
functions-framework
google-cloud-firestore
google-cloud-resource-manager
google-api-python-client
